NFTs
This contract extends the ERC721 and Ownable contracts from the OpenZeppelin library, which provide standard functionality for creating and managing NFTs.

The mint function allows the contract owner to mint a new NFT and assign it to a specified recipient address, along with a given metadata URI. The _tokenIdCounter variable is used to keep track of the current token ID, which is incremented each time a new NFT is minted.

Make sure to save the code in a .sol file and follow the steps mentioned in the previous answer to add it to your GitHub repository. Don't forget to include the necessary imports as well, which in this case are ERC721 and Ownable from the OpenZeppelin library.